Refactor code to add choaticbot package
