+++ src/main/java/choaticbot/inputs/InputProcessor.java
@@ -1,4 +1,4 @@
-package inputs;
+package choaticbot.inputs;
 
 public class InputProcessor {
