+++ src/main/java/choaticbot/tasks/Deadlines.java
@@ -1,4 +1,4 @@
-package tasks;
+package choaticbot.tasks;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
