+++ text-ui-test/input.txt
@@ -1,5 +1,5 @@
 todo todo1
-deadline deadline1 /by Sunday
+deadline deadline1 /by 2024-09-03 18:00
 event buy a kite event /from Tuesday /to Saturday
 list
 mark 1
